Clion's refactor tool didn't pick-up all these corrections when I renamed 'codegen/utils' to 'codegen/util' #690
