+++ src/codegen/aggregation.cpp
@@ -12,7 +12,7 @@
 
 #include "codegen/aggregation.h"
 
-#include "include/codegen/utils/if.h"
+#include "include/codegen/util/if.h"
 #include "codegen/type.h"
 #include "common/logger.h"
