+++ src/codegen/table.cpp
@@ -14,7 +14,7 @@
 
 #include "catalog/schema.h"
 #include "include/codegen/proxy/data_table_proxy.h"
-#include "include/codegen/utils/loop.h"
+#include "include/codegen/util/loop.h"
 #include "include/codegen/proxy/runtime_functions_proxy.h"
 #include "storage/data_table.h"
