+++ src/codegen/type.cpp
@@ -12,7 +12,7 @@
 
 #include "codegen/type.h"
 
-#include "include/codegen/utils/if.h"
+#include "include/codegen/util/if.h"
 #include "include/codegen/proxy/values_runtime_proxy.h"
 #include "type/value.h"
