+++ src/codegen/utils/cc_hash_table.cpp
@@ -4,13 +4,13 @@
 //
 // cc_hash_table.cpp
 //
-// Identification: src/codegen/utils/cc_hash_table.cpp
+// Identification: src/codegen/util/cc_hash_table.cpp
 //
 // Copyright (c) 2015-17, Carnegie Mellon University Database Group
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/utils/cc_hash_table.h"
+#include "codegen/util/cc_hash_table.h"
 
 #include "common/logger.h"
 #include "common/platform.h"
