+++ src/codegen/utils/oa_hash_table.cpp
@@ -4,13 +4,13 @@
 //
 // oa_hash_table.cpp
 //
-// Identification: src/codegen/utils/oa_hash_table.cpp
+// Identification: src/codegen/util/oa_hash_table.cpp
 //
 // Copyright (c) 2015-17, Carnegie Mellon University Database Group
 //
 //===----------------------------------------------------------------------===//
 
-#include "codegen/utils/oa_hash_table.h"
+#include "codegen/util/oa_hash_table.h"
 
 #include <string.h>
