+++ src/codegen/value.cpp
@@ -15,7 +15,7 @@
 #include <list>
 #include <queue>
 
-#include "include/codegen/utils/if.h"
+#include "include/codegen/util/if.h"
 #include "include/codegen/proxy/runtime_functions_proxy.h"
 #include "codegen/type.h"
