+++ src/include/codegen/varlen.h
@@ -13,7 +13,7 @@
 #pragma once
 
 #include "codegen/codegen.h"
-#include "include/codegen/utils/if.h"
+#include "include/codegen/util/if.h"
 #include "include/codegen/proxy/runtime_functions_proxy.h"
 #include "include/codegen/proxy/varlen_proxy.h"
