+++ src/include/codegen/vectorized_loop.h
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include "include/codegen/utils/loop.h"
+#include "include/codegen/util/loop.h"
 
 namespace peloton {
 namespace codegen {
