+++ test/codegen/if_test.cpp
@@ -13,8 +13,8 @@
 #include "common/harness.h"
 
 #include "codegen/function_builder.h"
-#include "include/codegen/utils/if.h"
-#include "include/codegen/utils/loop.h"
+#include "include/codegen/util/if.h"
+#include "include/codegen/util/loop.h"
 #include "include/codegen/proxy/runtime_functions_proxy.h"
 
 #include "codegen/codegen_test_util.h"
